requirements.txt
fastapi
uvicorn
pandas
networkx
python-louvain
google-generativeai
sentence-transformers
numpy
python-dotenv
aiofiles
